+++ crates/mpcs/src/lib.rs
@@ -270,10 +270,9 @@ pub use basefold::{
 pub mod jagged;
 pub use jagged::{
     JaggedBatchOpenProof, JaggedCommitment, JaggedCommitmentWithWitness, JaggedSumcheckInput,
-    jagged_batch_open, jagged_batch_verify, jagged_commit, jagged_sumcheck_prove,
+    evaluate_g, evaluate_g_backward, evaluate_g_forward, jagged_batch_open, jagged_batch_verify,
+    jagged_commit, jagged_sumcheck_prove,
 };
-pub mod jagged_evaluator;
-pub use jagged_evaluator::{evaluate_g, evaluate_g_backward, evaluate_g_forward};
 #[cfg(feature = "whir")]
 extern crate whir as whir_external;
 #[cfg(feature = "whir")]
